requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic<2
